+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ========
 
 * Fixed a couple issues creating a reverse lookup DNS zone in route 53
+* Updated the s3organizer command to better support re-encryption in the same bucket, and clarified prompts
 
 0.0.27
 ======